move some graphql definitions inside the `views` module
